modified

+++ app/Http/Controllers/JobController.php
@@ -3,90 +3,98 @@
 namespace App\Http\Controllers;
 
 use App\Models\Job;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+
+
 
 class JobController extends Controller
 {
-    public function index(){
-        
+    public function index()
+    {
+
         $jobs = Job::with('employer')->latest()->simplePaginate(3);
 
-             return view('jobs/index', [
-              
-                 'jobs' => $jobs,
+        return view('jobs/index', [
 
-     ]);
+            'jobs' => $jobs,
 
-
+        ]);
     }
 
-    public function create (){
+    public function create()
+    {
 
-    return view('jobs/create');
-
-        
+        return view('jobs/create');
     }
 
-    public function show(Job $job){
+    public function show(Job $job)
+    {
 
-    return view('jobs/show', ['job' => $job]);
-
+        return view('jobs/show', ['job' => $job]);
     }
 
-    public function store(){
+    public function store()
+    {
 
         request()->validate([
             'title' => ['required', 'min:3'],
             'salary' => ['required'],
         ]);
-    
+
         Job::create([
             'title' => request('title'),
-           'salary' => request('salary'),
-           'employer_id' => 1
-           
+            'salary' => request('salary'),
+            'employer_id' => 1
+
         ]);
-    
+
         return redirect('/jobs');
     }
 
-    public function edit(Job $job){
+    public function edit(Job $job)
+    {
 
-    return view('jobs.edit', ['job' => $job]);
+     
+      
+        Gate::authorize('edit-job', $job);
 
+  
+        return view('jobs.edit', ['job' => $job]);
     }
 
-    public function update(Job $job){
+    public function update(Job $job)
+    {
 
-         // authorization
-    
-    
-    // validate
-
-    request()->validate([
-        'title' => ['required', 'min:3'],
-        'salary' => ['required'],
-    ]);
-
-    // update job
-    
-    $job->update([
-        'title' => request('title'),
-        'salary' => request('salary'),
-    ]);
-    
+        Gate::authorize('edit-job', $job);
  
-    // redirect
-   
+        request()->validate([
+            'title' => ['required', 'min:3'],
+            'salary' => ['required'],
+        ]);
 
-    return redirect('/jobs/' . $job->id);
-   
+        // update job
+
+        $job->update([
+            'title' => request('title'),
+            'salary' => request('salary'),
+        ]);
+
+
+        // redirect
+
+
+        return redirect('/jobs/' . $job->id);
     }
 
-    public function destroy(Job $job){
+    public function destroy(Job $job)
+    {
+        Gate::authorize('edit-job', $job);
+
 
         $job->delete();
         return redirect('/jobs');
-
     }
 }
